David Luff:
Some links are broken.  I've fixed the SRTM data links, plus the one for
FGFLY (very dead project that I've pointed to the only reference I could
find left to it).

Curt:

I'd like to leave the main SRTM link as is since it points to the top level
(rather than pointing to a specific region.)